Isolate needless static page functionality into seperate controller.
Some of the code in the items controller is from a CI tutorial,
and the purpose of that code was to render static pages.

This is not the item controller's job. The item controller renders ITEMS, not static pages.

Static page rendering functionality has been separated into a new page controller